blocks other actions in OS visualization for coordenador

+++ resources/views/operacional/os/index.blade.php
@@ -26,14 +26,18 @@
 
                     <th class="features">{{ $agenda->cliente->nome }} {{ $agenda->data }}</th>
                     <td><a href="{{ action('AmbientesController@edit', ['os' => $agenda->os]) }}"><img src="{{ asset('images/icons/testimonial32.png') }}" alt="O.S."/></a></td>
-                    <td><a href="{{ action('AgendaController@edit', [$agenda]) }}"><img src="{{ asset('images/icons/moleskine32.png') }}" alt="Agenda"/></a></td>
-                    <td><a href="{{ action('OsController@show', $agenda->os) }}"><img src="{{ asset('images/icons/search.png') }}" alt="Visualizar"/></a></td>
-                    <td><a href="{{ action('OsController@edit', $agenda->os) }}"><img src="{{ asset('images/icons/pencil32.png') }}" alt="Editar"/></a></td>
-                    <td>
-                        {!! Form::open(['action' => ['OsController@destroy', $agenda->os], 'method' => 'delete']) !!}
-                        <button class="small button red delete" style="border: none">Excluir</button>
-                        {!! Form::close() !!}
-                    </td>
+                    @can('gerente')
+                        <td><a href="{{ action('AgendaController@edit', [$agenda]) }}"><img src="{{ asset('images/icons/moleskine32.png') }}" alt="Agenda"/></a></td>
+                        <td><a href="{{ action('OsController@show', $agenda->os) }}"><img src="{{ asset('images/icons/search.png') }}" alt="Visualizar"/></a></td>
+                        <td><a href="{{ action('OsController@edit', $agenda->os) }}"><img src="{{ asset('images/icons/pencil32.png') }}" alt="Editar"/></a></td>
+                        <td>
+                            {!! Form::open(['action' => ['OsController@destroy', $agenda->os], 'method' => 'delete']) !!}
+                            <button class="small button red delete" style="border: none">Excluir</button>
+                            {!! Form::close() !!}
+                        </td>
+                    @else
+                        <td colspan="4"></td>
+                    @endcan
                 </tr>
             </tbody>
             @endforeach
